量化分值

+++ svision Maven Webapp/src/main/java/com/rmbank/supervision/model/ItemProcessGrade.java
@@ -7,7 +7,7 @@ public class ItemProcessGrade {
 
     private Integer gradeDetailId;
 
-    private Integer grade;
+    private Double grade;
 
     public Integer getId() {
         return id;
@@ -33,11 +33,11 @@ public class ItemProcessGrade {
         this.gradeDetailId = gradeDetailId;
     }
 
-    public Integer getGrade() {
+    public Double getGrade() {
         return grade;
     }
 
-    public void setGrade(Integer grade) {
+    public void setGrade(Double grade) {
         this.grade = grade;
     }
 }
